+++ components/Layout.tsx
@@ -11,7 +11,6 @@ import UserMenu from "./UserMenu";
 import React, { useEffect, useState } from "react";
 import useListConversations from "../hooks/useListConversations";
 import { useXmtpStore } from "../store/xmtp";
-import useInitXmtpClient from "../hooks/useInitXmtpClient";
 import useWindowSize from "../hooks/useWindowSize";
 
 const Layout: React.FC<{ children?: React.ReactNode }> = ({ children }) => {
@@ -33,6 +32,52 @@ const Layout: React.FC<{ children?: React.ReactNode }> = ({ children }) => {
     watchAccount(() => resetXmtpState());
   }, []);
 
+  const conversationView = () => {
+    return (
+      <ConversationView>
+        {walletAddress && client ? (
+          <>
+            <div className="flex bg-zinc-50 border-b border-gray-200 md:bg-white md:border-0 max-h-16 min-h-[4rem]">
+              <RecipientControl setShowMessageView={setShowMessageView} />
+            </div>
+            {children}
+          </>
+        ) : (
+          <XmtpInfoPanel />
+        )}
+      </ConversationView>
+    );
+  };
+
+  const navigationView = () => {
+    return (
+      <NavigationView>
+        <aside className="flex w-full md:w-84 flex-col flex-grow fixed inset-y-0">
+          <div className="flex flex-col flex-grow md:border-r md:border-gray-200 bg-white overflow-y-auto">
+            <div className="max-h-16 min-h-[4rem] bg-p-600 flex items-center justify-between flex-shrink-0 px-4">
+              <Link href="/" passHref={true}>
+                <img
+                  className="h-8 w-auto"
+                  src="/xmtp-icon.png"
+                  alt="XMTP"
+                  data-testid="xmtp-logo"
+                />
+              </Link>
+              {walletAddress && client && (
+                <NewMessageButton setShowMessageView={setShowMessageView} />
+              )}
+            </div>
+            <NavigationPanel isError={!!error} />
+            <UserMenu
+              isError={!!error}
+              setShowMessageView={setShowMessageView}
+            />
+          </div>
+        </aside>
+      </NavigationView>
+    );
+  };
+
   return (
     <>
       <Head>
@@ -45,83 +90,13 @@ const Layout: React.FC<{ children?: React.ReactNode }> = ({ children }) => {
       <div>
         {size[0] > 600 ? (
           <>
-            <NavigationView>
-              <aside className="flex w-full md:w-84 flex-col flex-grow fixed inset-y-0">
-                <div className="flex flex-col flex-grow md:border-r md:border-gray-200 bg-white overflow-y-auto">
-                  <div className="max-h-16 min-h-[4rem] bg-p-600 flex items-center justify-between flex-shrink-0 px-4">
-                    <Link href="/" passHref={true}>
-                      <img
-                        className="h-8 w-auto"
-                        src="/xmtp-icon.png"
-                        alt="XMTP"
-                        data-testid="xmtp-logo"
-                      />
-                    </Link>
-                    {walletAddress && client && (
-                      <NewMessageButton
-                        setShowMessageView={setShowMessageView}
-                      />
-                    )}
-                  </div>
-                  {<NavigationPanel isError={!!error} />}
-                  <UserMenu
-                    isError={!!error}
-                    setShowMessageView={setShowMessageView}
-                  />
-                </div>
-              </aside>
-            </NavigationView>
-            <ConversationView>
-              {walletAddress && client ? (
-                <>
-                  <div className="flex bg-zinc-50 border-b border-gray-200 md:bg-white md:border-0 max-h-16 min-h-[4rem]">
-                    <RecipientControl setShowMessageView={setShowMessageView} />
-                  </div>
-                  {children}
-                </>
-              ) : (
-                <XmtpInfoPanel />
-              )}
-            </ConversationView>
+            {navigationView()}
+            {conversationView()}
           </>
         ) : showMessageView || recipientWalletAddress ? (
-          <ConversationView>
-            {walletAddress && client ? (
-              <>
-                <div className="flex bg-zinc-50 border-b border-gray-200 md:bg-white md:border-0 max-h-16 min-h-[4rem]">
-                  <RecipientControl setShowMessageView={setShowMessageView} />
-                </div>
-                {children}
-              </>
-            ) : (
-              <XmtpInfoPanel />
-            )}
-          </ConversationView>
+          <>{conversationView()}</>
         ) : (
-          <NavigationView>
-            <aside className="flex w-full md:w-84 flex-col flex-grow fixed inset-y-0">
-              <div className="flex flex-col flex-grow md:border-r md:border-gray-200 bg-white overflow-y-auto">
-                <div className="max-h-16 min-h-[4rem] bg-p-600 flex items-center justify-between flex-shrink-0 px-4">
-                  <Link href="/" passHref={true}>
-                    <img
-                      className="h-8 w-auto"
-                      src="/xmtp-icon.png"
-                      alt="XMTP"
-                      data-testid="xmtp-logo"
-                    />
-                  </Link>
-                  {walletAddress && client && (
-                    <NewMessageButton setShowMessageView={setShowMessageView} />
-                  )}
-                </div>
-                {<NavigationPanel isError={!!error} />}
-                <UserMenu
-                  isError={!!error}
-                  setShowMessageView={setShowMessageView}
-                />
-              </div>
-            </aside>
-          </NavigationView>
+          <> {navigationView()}</>
         )}
       </div>
     </>
